Implement eBay driver that can search products by keywords
The eBayPanhandler class implements what is currently the sole
function of the Panhandles interface: get_products_by_keywords().  It
will return an array of PanhandlerProduct objects representing the
products in the search results for the given keywords.

Being an initial version, the driver has some limitations:

    1. It does not provide a way to limit the results.  Only the first
    hundred will be returned.

    2. It does not provide a means of sorting the results.

    3. It does not expose any of the pagination features that eBay
    offers for their search results.

    4. Probably a lot of other things.

However, it is usable as-is.

A demonstration of driver can be found in the Tests directory.  Simply
run the test script from the command line and you will see list of a
bunch of Love Hina paraphernalia (because I have been watching that
show almost non-stop at home, so I have it on the brain).