CJ: Implement set_results_page()
As noted in the comments for the code, the documentation for
Commission Junction claims that the first page of results is numbered
at zero.  But this doesn’t appear to be the case.  If we try setting
‘page-number’ in the request to be zero then we do not get any
results.  So the code makes the assumption that the documentation is
incorrect and that the first page begins with one.

Redmine-Ticket: 2057