eBay: Return a more meaningful description for products
The eBay API does not return to us a single field with a user-given
description of the product.  We could scrape the product page to get
that information, but right now that’s a huge headache.  So instead we
take some information that most consumers would want to know and put
that into a list, which we then return as the description.