eBay: Allow fetching products from specific sellers
Users may want to not only display products based on certain keywords,
but also restrict the products to a certain seller or list of sellers.
This can be useful, for instance, if the driver is being used to show
products as part of an online store, where the site owner wants only
his products to be displayed.

This commit teaches get_products_by_keywords() to look for ‘sellers’
in the $options array.  When given, products are restricted to those
associated with those seller names.  The new apply_filter() method
adds the seller names to the request URL.  However, it currently has a
flaw.  It assumes that there are no other filters, and so it sets the
seller as ‘itemFilters(0)’.  Ideally apply_filters() should count how
many filters are being applied so that it can correctly have
‘itemFilters(N)’ for ‘N’ number of filters.  Since we only support
filtering by sellers right now, the method lazily sets the seller name
as the first filler.

This is also a new method for determining whether or not an XML
response from eBay indicates a successful search result.  This helps
stop extract_products() from looping over non-existent items when
there is in an error with the request.

Redmine-Ticket: 2117