Add an exception to be thrown when missing requirements
Many of the drivers will rely on functionality from other third-party
libraries in order to function properly.  If they are unable to load
those libraries then they can throw this exception.  This provides the
chance to take some action such as installing the requirement, or
disabling the use of the driver as a whole.  This exception should be
thrown in favor of calling die() in such circumstances, because the
drivers should not entirely halt the application using them.

Redmine-Ticket: 2112