eBay: Implement get_products_from_vendor()
Much of the code we use to get all of the products for a specific
vendor is the same code we use to get products matching certain
keywords.  So this gives us a chance to refactor the driver.

First we move $keywords into a private member of the class, and the
existing get_products_by_keywords() method saves its argument in that
member.  We make this change so that we can remove the $keyword
argument from make_request_url().  That method is now more generic,
and now calls ‘findItemsAdvanced’ through eBay instead of
‘findItemsByKeywords’.  Other functions also lost $keyword as an
argument so that they could be more generic and their logic could be
recycled for finding products based on other criteria.

The new make_request_url() looks for the presence of either vendor IDs
or keywords, and uses those to put together the appropriate parameters
for ‘findItemsAdvanced’.  Since we can also search by keywords using
‘findItemsAdvanced’, we can get away with a small amount of changes to
support this.

Redmine-Ticket: 2123