eBay: Refactor the parsing of $options in interface methods
The logic for dealing with $options is very similar in
get_products_from_vendor() and get_products_by_keywords().  As new
options are added, that code duplication will only grow.

Here we factor out that logic into a private method which is
responsible for setting appropriate class members based on the values
in $options.  The interface methods then call this new method before
doing anything else.

This makes it possible to change options without having to necessarily
change logic in all of the public functions.