Add get_products and get_supported_options to panhandle interface
get_products effectively replaces get_products_by_keywords and
get_products_from_vendor. The reasoning for this is that the interface
defines a more generic method path for things such as the plugin class
to call, the drivers themselves will then be responsible for parsing
the information passed in. This does not prevent extra
get_product_yadda methods from existing within the drivers.

get_supported_options() returns an array list of supported attributes
that the driver is capable of understanding. The purpose of this is to
filter out erroneous commands that may be mistakenly put into a
shortcode or other input before they are sent to get_products().