Provide a method for setting default driver option values
There is currently no good way to specify default values for driver
options.  The driver must have access to a list of all of the options
it accepts, because it is required to return them for
get_supported_options().  So initially it may seem plausible to put
default option values in the driver definition itself.

However, this approach fails for situations where the default values
need to change based on the use of the driver.  For instance, for the
MoneyPress plugins the default values used for eBay or Commission
Junction can change from one installation to another (e.g. seller
IDs).  In those situations it is not acceptable to have the defaults
written directly in the driver, because that would require code
changes for each installation.

We can avoid this problem by requiring all drivers to accept a hash of
default driver values, which they can then save and rely on when those
options are not given to get_products().  This will give us a
consistent way of passing in defaults on a per-use basis without
having to modify the code of the drivers themselves.

Redmine-Ticket: 2196
Redmine-Ticket: 2261