eBay: Ensure that $keywords is an array
The WPSCL-Generic plugin will pass in the ‘keywords’ short-code
parameter as a string, but the driver expects this to be an array.
Originally this was done in the MoneyPress eBay plugin itself, but now
we have to break up the keyword string into an array here in the
driver before fetching any product information.