Too many updates for a single commit. :(
1) Added the pieces to get the Amazon product array
setup.  A lot of stuff to process the images out of
the XML, create a basic description (aka Editorial Review)
and do pricing lookups.

2) Added various helper functions for the above item.

3) Took out some hard-coded driver defaults that was
used to early development, they are now passed in as
standard attributes or driver settings.

4) Made the amazon request parameters a map of
the CamelCase amazon keywords to lowercase WPCSL
generic driver attributes so we can map them on the
fly before calling Amazon.

I know... break up the commits.